Refactor `ayon_style.py` to `style.py`
