+++ client/ayon_ui_qt/components/entity_path.py
@@ -6,7 +6,7 @@ from typing import Optional
 from PySide6.QtCore import Qt
 from qtpy.QtWidgets import QWidget
 
-from ..ayon_style import get_ayon_style
+from ..style import get_ayon_style
 from ..utils import clear_layout
 from .label import AYLabel
 from .layouts import AYHBoxLayout
