+++ client/ayon_ui_qt/components/filter.py
@@ -18,7 +18,7 @@ from qtpy.QtCore import QModelIndex, QPersistentModelIndex, Qt, Signal
 from qtpy.QtGui import QColor, QPainter
 from qtpy.QtWidgets import QStyle, QStyleOptionViewItem
 
-from ..ayon_style import get_ayon_style
+from ..style import get_ayon_style
 from ..variants import QFrameVariants, QStyledItemDelegateVariants
 
 from .buttons import AYButton
