+++ client/ayon_ui_qt/components/label.py
@@ -20,7 +20,7 @@ try:
 except ImportError:
     from ..vendor.qtmaterialsymbols import get_icon
 
-from ..ayon_style import get_ayon_style
+from ..style import get_ayon_style
 from ..color_utils import compute_color_for_contrast
 from ..variants import QLabelVariants
