+++ client/ayon_ui_qt/tester.py
@@ -4,8 +4,8 @@ from pathlib import Path
 
 from qtpy import QtWidgets
 
-# from .ayon_style import AYONStyle
-from .ayon_style import get_ayon_style
+# from .style import AYONStyle
+from .style import get_ayon_style
 
 
 class Style(Enum):
